+++ CoachBot/Controllers/TournamentGroupController.cs
@@ -1,10 +1,9 @@
 ﻿using CoachBot.Domain.Model;
 using CoachBot.Domain.Services;
+using CoachBot.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace CoachBot.Controllers
 {
@@ -14,39 +13,66 @@ namespace CoachBot.Controllers
     public class TournamentGroupController : Controller
     {
         private readonly TournamentService _tournamentService;
+        private readonly PlayerService _playerService;
 
-        public TournamentGroupController(TournamentService tournamentService)
+        public TournamentGroupController(TournamentService tournamentService, PlayerService playerService)
         {
             _tournamentService = tournamentService;
+            _playerService = playerService;
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public void DeleteTournamentGroup(int id)
         {
+            if (!_playerService.IsAdmin(User.GetSteamId()))
+            {
+                throw new Exception();
+            }
             _tournamentService.DeleteTournamentGroup(id);
         }
 
+        [Authorize]
         [HttpPost]
         public void CreateTournamentGroup(TournamentGroup tournamentGroup)
         {
+            if (!_playerService.IsAdmin(User.GetSteamId()))
+            {
+                throw new Exception();
+            }
             _tournamentService.CreateTournamentGroup(tournamentGroup);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public void UpdateTournamentGroup(TournamentGroup tournamentGroup)
         {
+            if (!_playerService.IsAdmin(User.GetSteamId()))
+            {
+                throw new Exception();
+            }
             _tournamentService.UpdateTournamentGroup(tournamentGroup);
         }
 
+        [Authorize]
         [HttpPost("{id}/teams")]
         public void AddTournamentGroupTeam(TournamentGroupTeamDto tournamentGroupTeamDto)
         {
+            if (!_playerService.IsAdmin(User.GetSteamId()))
+            {
+                throw new Exception();
+            }
             _tournamentService.AddTournamentTeam(tournamentGroupTeamDto.TeamId, tournamentGroupTeamDto.TournamentGroupId);
         }
 
+        [Authorize]
         [HttpPost("{id}/teams/{teamId}")]
         public void RemoveTournamentGroupTeam(int id, int teamId)
         {
+            if (!_playerService.IsAdmin(User.GetSteamId()))
+            {
+                throw new Exception();
+            }
             _tournamentService.RemoveTournamentTeam(teamId, id);
         }
     }
